@@ -1,30 +1,21 @@
-buildscript {
-    repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.+'
-    }
-}
 apply plugin: 'com.android.application'
 apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "21.1.1"
+    compileSdkVersion 27
+    buildToolsVersion "27.0.3"
 
     def versionPropsFile = file('version.properties')
 
     if (versionPropsFile.canRead()) {
-        def Properties versionProps = new Properties()
+        Properties versionProps = new Properties()
 
         versionProps.load(new FileInputStream(versionPropsFile))
         def value = 0
         def runTasks = gradle.startParameter.taskNames
         println runTasks
         if (':app:assembleRelease' in runTasks) {
-            value = 1;
+            value = 1
         }
 
         def versionMajor = 1
@@ -44,7 +35,7 @@ android {
             versionCode version_Code
             versionName "${versionMajor}.${versionMinor}.${versionPatch} (${versionBuild})"
             minSdkVersion 18
-            targetSdkVersion 19
+            targetSdkVersion 27
         }
     } else {
         throw new GradleException("Could not read version.properties!")
@@ -56,19 +47,24 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
     }
-}
 
-repositories {
-    mavenCentral()
-    maven { url 'https://maven.fabric.io/public' }
+    dataBinding {
+        enabled = true
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    compile 'com.android.support:support-v4:19.1.0'
-    compile 'me.grantland:autofittextview:0.2.+'
-    compile 'com.github.lecho:hellocharts-library:1.5.8@aar'
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile('com.crashlytics.sdk.android:crashlytics:2.5.7@aar') {
-        transitive = true;
+    implementation 'com.android.support:support-v4:27.1.0'
+    implementation 'com.android.support:appcompat-v7:27.1.0'
+    implementation 'com.github.lecho:hellocharts-library:1.5.8@aar'
+    implementation 'com.jakewharton.timber:timber:4.6.1'
+    implementation('com.crashlytics.sdk.android:crashlytics:2.9.1@aar') {
+        transitive = true
     }
+    implementation 'com.google.firebase:firebase-core:11.8.0'
 }
