@@ -1,26 +1,24 @@
 
 package com.mooshim.mooshimeter.activities;
 
-import android.app.ActionBar;
 import android.content.Context;
 import android.content.Intent;
 import android.content.res.Configuration;
+import android.databinding.DataBindingUtil;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.support.v7.app.ActionBar;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
 import android.text.style.RelativeSizeSpan;
-import android.util.Log;
 import android.util.TypedValue;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.Button;
-import android.widget.ImageView;
 import android.widget.PopupMenu;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -29,37 +27,40 @@ import com.mooshim.mooshimeter.R;
 import com.mooshim.mooshimeter.common.BroadcastIntentData;
 import com.mooshim.mooshimeter.common.CooldownTimer;
 import com.mooshim.mooshimeter.common.MeterReading;
-import com.mooshim.mooshimeter.interfaces.MooshimeterControlInterface;
-import com.mooshim.mooshimeter.interfaces.MooshimeterDelegate;
-import com.mooshim.mooshimeter.devices.MooshimeterDeviceBase;
-import com.mooshim.mooshimeter.interfaces.NotifyHandler;
 import com.mooshim.mooshimeter.common.SpeaksOnLargeChange;
 import com.mooshim.mooshimeter.common.Util;
+import com.mooshim.mooshimeter.databinding.ActivityMeterBinding;
+import com.mooshim.mooshimeter.databinding.ElementDeviceActivityTitlebarBinding;
+import com.mooshim.mooshimeter.devices.MooshimeterDeviceBase;
+import com.mooshim.mooshimeter.interfaces.MooshimeterControlInterface;
+import com.mooshim.mooshimeter.interfaces.MooshimeterDelegate;
+import com.mooshim.mooshimeter.interfaces.NotifyHandler;
 
 import java.util.List;
 
-import me.grantland.widget.AutofitHelper;
+import timber.log.Timber;
 
-public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
-    private static final String TAG = "DeviceActivity";
+public class DeviceActivity extends BaseActivity implements MooshimeterDelegate {
     public static final String AUTORANGE = "AUTORANGE";
+    private ActivityMeterBinding viewBinding;
+    private ElementDeviceActivityTitlebarBinding titlebarBinding;
 
     private static MooshimeterControlInterface.Channel chanEnum(int c) {
         return MooshimeterControlInterface.Channel.values()[c];
     }
 
 
-	// BLE
+    // BLE
     private MooshimeterDeviceBase mMeter;
 
     // GUI
     private final TextView[] value_labels = new TextView[2];
     private TextView power_label;
 
-    private static final Button[] input_set_buttons = {null,null};
-    private static final Button[] range_buttons     = {null,null};
-    private static final Button[] zero_buttons      = {null,null};
-    private static final Button[] sound_buttons     = {null,null};
+    private static final Button[] input_set_buttons = {null, null};
+    private static final Button[] range_buttons = {null, null};
+    private static final Button[] zero_buttons = {null, null};
+    private static final Button[] sound_buttons = {null, null};
 
     private Button rate_button;
     private Button depth_button;
@@ -70,9 +71,10 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
     private float battery_voltage = 0;
 
     // GUI housekeeping
-    private Drawable getAutoBG(){
+    private Drawable getAutoBG() {
         return getResources().getDrawable(R.drawable.button_auto);
     }
+
     private Drawable getNormalBG() {
         return getResources().getDrawable(R.drawable.button_normal);
     }
@@ -82,121 +84,109 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
     private SpeaksOnLargeChange speaksOnLargeChange = new SpeaksOnLargeChange();
     private BroadcastIntentData broadcastIntentData = new BroadcastIntentData(); // added for broadcast intent
 
-    private TextView findAndAutofit(int id) {
-        TextView rval = (TextView)findViewById(id);
-        AutofitHelper a = AutofitHelper.create(rval);
-        a.setMaxLines(1);
-        a.setEnabled(true);
-        return rval;
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
+        super.onCreate(savedInstanceState);
+
+        viewBinding = DataBindingUtil.setContentView(this, R.layout.activity_meter);
+
+        // Bind the GUI elements
+        value_labels[0] = viewBinding.ch1ValueLabel;
+        value_labels[1] = viewBinding.ch2ValueLabel;
+        power_label = viewBinding.powerLabel;
+
+        input_set_buttons[0] = viewBinding.ch1InputSetButton;
+        range_buttons[0] = viewBinding.ch1RangeButton;
+        zero_buttons[0] = viewBinding.ch1ZeroButton;
+        sound_buttons[0] = viewBinding.ch1SoundButton;
+
+        input_set_buttons[1] = viewBinding.ch2InputSetButton;
+        range_buttons[1] = viewBinding.ch2RangeButton;
+        zero_buttons[1] = viewBinding.ch2ZeroButton;
+        sound_buttons[1] = viewBinding.ch2SoundButton;
+
+        rate_button = viewBinding.rateButton;
+        depth_button = viewBinding.depthButton;
+        logging_button = viewBinding.logButton;
+        graph_button = viewBinding.graphButton;
+        power_button = viewBinding.powerButton;
+
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayShowHomeEnabled(true);
+            actionBar.setDisplayShowTitleEnabled(false);
+            titlebarBinding = ElementDeviceActivityTitlebarBinding.inflate(getLayoutInflater());
+            actionBar.setCustomView(titlebarBinding.getRoot());
+            actionBar.setDisplayShowCustomEnabled(true);
+        }
     }
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-		requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
-		super.onCreate(savedInstanceState);
-
-        // GUI
-        setContentView(R.layout.activity_meter);
-
-        // Bind the GUI elements
-        value_labels[0]        =         findAndAutofit(R.id.ch1_value_label);
-        value_labels[1]        =         findAndAutofit(R.id.ch2_value_label);
-        power_label            =         findAndAutofit(R.id.power_label);
-
-        input_set_buttons  [0] = (Button)findAndAutofit(R.id.ch1_input_set_button);
-        range_buttons      [0] = (Button)findViewById(R.id.ch1_range_button);
-        zero_buttons       [0] = (Button)findAndAutofit(R.id.ch1_zero_button);
-        sound_buttons      [0] = (Button)findAndAutofit(R.id.ch1_sound_button);
-
-        input_set_buttons  [1] = (Button)findAndAutofit(R.id.ch2_input_set_button);
-        range_buttons      [1] = (Button)findViewById(R.id.ch2_range_button);
-        zero_buttons       [1] = (Button)findAndAutofit(R.id.ch2_zero_button);
-        sound_buttons      [1] = (Button)findAndAutofit(R.id.ch2_sound_button);
-
-        rate_button            = (Button)findViewById(R.id.rate_button);
-        depth_button           = (Button)findViewById(R.id.depth_button);
-        logging_button         = (Button)findAndAutofit(R.id.log_button);
-        graph_button           = (Button)findAndAutofit(R.id.graph_button);
-        power_button           = (Button)findAndAutofit(R.id.power_button);
-
-        ActionBar actionBar = getActionBar();
-        actionBar.setDisplayShowHomeEnabled(true);
-        actionBar.setDisplayShowTitleEnabled(false);
-        actionBar.setCustomView(R.layout.element_device_activity_titlebar);
-        actionBar.setDisplayShowCustomEnabled(true);
-	}
-
-	@Override
-	public void onDestroy() {
-		super.onDestroy();
+    public void onDestroy() {
+        super.onDestroy();
         setResult(RESULT_OK);
-	}
+    }
 
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		//this.optionsMenu = menu;
-		// Inflate the menu items for use in the action bar
-		MenuInflater inflater = getMenuInflater();
-		inflater.inflate(R.menu.device_activity_actions, menu);
-		return super.onCreateOptionsMenu(menu);
-	}
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        //this.optionsMenu = menu;
+        // Inflate the menu items for use in the action bar
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.device_activity_actions, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
 
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		// Handle presses on the action bar items
-		switch (item.getItemId()) {
-		case R.id.opt_prefs:
-            pushActivityToStack(mMeter,PreferencesActivity.class);
-			break;
-		default:
-            finish();
-            return true;
-		}
-		return true;
-	}
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle presses on the action bar items
+        switch (item.getItemId()) {
+            case R.id.opt_prefs:
+                pushActivityToStack(mMeter, PreferencesActivity.class);
+                break;
+            default:
+                finish();
+                return true;
+        }
+        return true;
+    }
 
     @Override
     protected void onPause() {
         super.onPause();
         getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         // I still don't understand why, but mMeter keeps getting set to null!  WHY ANDROID WHY?!
-        if(mMeter==null) {
-            Log.e(TAG,"GOT A NULL MMETER IN onPause!");
+        if (mMeter == null) {
+            Timber.e("GOT A NULL MMETER IN onPause!");
             return;
         }
-        if(!(mMeter.speech_on.get(MooshimeterControlInterface.Channel.CH1) || mMeter.speech_on.get(MooshimeterControlInterface.Channel.CH2))) {
-            Util.dispatch(new Runnable() {
-                @Override
-                public void run() {
-                    mMeter.pause();
-                    mMeter.removeDelegate();
-                }
+        if (!(mMeter.speech_on.get(MooshimeterControlInterface.Channel.CH1) || mMeter.speech_on.get(MooshimeterControlInterface.Channel.CH2))) {
+            Util.dispatch(() -> {
+                mMeter.pause();
+                mMeter.removeDelegate();
             });
         }
     }
 
-	@Override
-	protected void onResume() {
-		super.onResume();
+    @Override
+    protected void onResume() {
+        super.onResume();
         getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         final MooshimeterDelegate d = this;
 
         // When resuming, Android may have destroyed objects we care about
         // Just fall back to the scan screen...
-        if(   mMeter==null
-           ||!mMeter.isConnected()) {
+        if (mMeter == null
+                || !mMeter.isConnected()) {
             onBackPressed();
             return;
         }
 
-        Util.dispatch(new Runnable() {
-            @Override
-            public void run() {
-                mMeter.addDelegate(d);
-                mMeter.stream();
-                Log.i(TAG, "Stream requested");
-                refreshAllControls();
-            }
+        Util.dispatch(() -> {
+            mMeter.addDelegate(d);
+            mMeter.stream();
+            Timber.i("Stream requested");
+            refreshAllControls();
         });
     }
 
@@ -207,12 +197,12 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
         // Double check here... still haven't figured this bug out.  FIXME
         Intent intent = getIntent();
         try {
-            mMeter = (MooshimeterDeviceBase)getDeviceWithAddress(intent.getStringExtra("addr"));
-        }catch(ClassCastException e) {
-            Log.e(TAG, "Squirrelly ClassCastException! Diagnose me!");
+            mMeter = (MooshimeterDeviceBase) getDeviceWithAddress(intent.getStringExtra("addr"));
+        } catch (ClassCastException e) {
+            Timber.e("Squirrelly ClassCastException! Diagnose me!");
             mMeter = null;
         }
-        if(mMeter==null) {
+        if (mMeter == null) {
             onBackPressed();
         }
     }
@@ -222,61 +212,52 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
         super.onConfigurationChanged(newConfig);
     }
 
-	private void setError(final String txt) {
+    private void setError(final String txt) {
         final Context c = this;
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                Toast.makeText(c, txt, Toast.LENGTH_LONG).show();
-            }
-        });
-	}
+        runOnUiThread(() -> Toast.makeText(c, txt, Toast.LENGTH_LONG).show());
+    }
 
     /////////////////////////
     // Widget Refreshers
     /////////////////////////
 
     private void refreshAllControls() {
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                rateButtonRefresh();
-                depthButtonRefresh();
-                loggingButtonRefresh();
-                graphButtonRefresh();
-                for (int c = 0; c < 2; c++) {
-                    inputSetButtonRefresh(c);
-                    rangeButtonRefresh(c);
-                    soundButtonRefresh(c);
-                    zeroButtonRefresh(c, mMeter.getOffset(chanEnum(c)));
-                }
+        runOnUiThread(() -> {
+            rateButtonRefresh();
+            depthButtonRefresh();
+            loggingButtonRefresh();
+            graphButtonRefresh();
+            for (int c = 0; c < 2; c++) {
+                inputSetButtonRefresh(c);
+                rangeButtonRefresh(c);
+                soundButtonRefresh(c);
+                zeroButtonRefresh(c, mMeter.getOffset(chanEnum(c)));
             }
         });
     }
+
     private void disableableButtonRefresh(Button b, String title, boolean en) {
-        final Drawable bg = en? getNormalBG(): getAutoBG();
+        final Drawable bg = en ? getNormalBG() : getAutoBG();
         b.setText(title);
         b.setBackground(bg);
     }
+
     private void autoButtonRefresh(final Button b, final String title, boolean auto) {
-        final Drawable bg = auto? getAutoBG(): getNormalBG();
+        final Drawable bg = auto ? getAutoBG() : getNormalBG();
         SpannableStringBuilder sb = new SpannableStringBuilder();
 
-        String auto_string = auto?"AUTO":"MANUAL";
+        String auto_string = auto ? "AUTO" : "MANUAL";
         sb.append(auto_string);
         sb.append("\n");
         int i = sb.length();
         sb.append(title);
-        sb.setSpan(new RelativeSizeSpan((float)1.6), i, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+        sb.setSpan(new RelativeSizeSpan((float) 1.6), i, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
 
         Util.setText(b, sb);
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                float button_height = b.getHeight(); // Button height in raw pixels
-                b.setTextSize(TypedValue.COMPLEX_UNIT_PX,button_height/3); // Divisor arrived at empirically
-                b.setBackground(bg);
-            }
+        runOnUiThread(() -> {
+            float button_height = b.getHeight(); // Button height in raw pixels
+            b.setTextSize(TypedValue.COMPLEX_UNIT_PX, button_height / 3); // Divisor arrived at empirically
+            b.setBackground(bg);
         });
     }
 
@@ -287,88 +268,82 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
 
     private void refreshTitle() {
         // Approximate remaining charge
-        double soc_percent = (battery_voltage - 2.0)*100.0;
+        double soc_percent = (battery_voltage - 2.0) * 100.0;
         soc_percent = Math.max(0, soc_percent);
-        soc_percent = Math.min(100,soc_percent);
-        final Drawable bat_img = getDrawableByURI("drawable/bat_icon_"+Integer.toString((int)soc_percent));
+        soc_percent = Math.min(100, soc_percent);
+        final Drawable bat_img = getDrawableByURI("drawable/bat_icon_" + Integer.toString((int) soc_percent));
 
         int rssi_val = mMeter.getRSSI();
         // rssi is always negative, map it to a percentage.
         // Let's just make -100db be 0%, and 0db be 100%
         rssi_val += 100;
         rssi_val = Math.max(0, rssi_val);
-        rssi_val = Math.min(100,rssi_val);
-        final Drawable rssi_img = getDrawableByURI("drawable/sig_icon_"+Integer.toString(rssi_val));
+        rssi_val = Math.min(100, rssi_val);
+        final Drawable rssi_img = getDrawableByURI("drawable/sig_icon_" + Integer.toString(rssi_val));
 
-        ActionBar ab = getActionBar();
-        final ViewGroup vg = (ViewGroup)ab.getCustomView();
-        if(vg==null) {
-            return;
-        }
-        final TextView title = (TextView)vg.findViewById(R.id.title_textview);
-        final ImageView bat  = (ImageView)vg.findViewById(R.id.bat_stat_img);
-        final ImageView rssi = (ImageView)vg.findViewById(R.id.rssi_img);
-
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                title.setText(mMeter.getBLEDevice().getName());
-                bat.setImageDrawable(bat_img);
-                rssi.setImageDrawable(rssi_img);
-            }
+        runOnUiThread(() -> {
+            titlebarBinding.titleTextview.setText(mMeter.getBLEDevice().getName());
+            titlebarBinding.batStatImg.setImageDrawable(bat_img);
+            titlebarBinding.rssiImg.setImageDrawable(rssi_img);
         });
     }
+
     private void mathLabelRefresh(final MeterReading val) {
-        Util.setText(power_label,val.toString());
+        Util.setText(power_label, val.toString());
         Util.setText(power_button, mMeter.getSelectedDescriptor(MooshimeterControlInterface.Channel.MATH).name);
     }
+
     private void mathButtonRefresh() {
         mathLabelRefresh(mMeter.getValue(MooshimeterControlInterface.Channel.MATH));
     }
-    private void graphButtonRefresh() {}
+
+    private void graphButtonRefresh() {
+    }
+
     private void rateButtonRefresh() {
         int rate = mMeter.getSampleRateHz();
-        String title = String.format("%dHz", rate);
+        String title = getString(R.string.rate_hz, rate);
         autoButtonRefresh(rate_button, title, mMeter.getRateAuto());
     }
+
     private void depthButtonRefresh() {
         int depth = mMeter.getBufferDepth();
-        String title = String.format("%dsmpl", depth);
+        String title = getString(R.string.depth_sample, depth);
         autoButtonRefresh(depth_button, title, mMeter.getDepthAuto());
     }
+
     private void loggingButtonRefresh() {
         int s = mMeter.getLoggingStatus();
         final String title;
-        final boolean logging_ok = s==0;
-        if(logging_ok) {
-            title = mMeter.getLoggingOn()?"Logging:ON":"Logging:OFF";
+        final boolean logging_ok = s == 0;
+        if (logging_ok) {
+            title = mMeter.getLoggingOn() ? "Logging:ON" : "Logging:OFF";
         } else {
             title = mMeter.getLoggingStatusMessage();
         }
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                disableableButtonRefresh(logging_button, title, logging_ok);
-            }
-        });
+        runOnUiThread(() -> disableableButtonRefresh(logging_button, title, logging_ok));
     }
+
     private void inputSetButtonRefresh(final int c) {
         final String s = mMeter.getInputLabel(chanEnum(c));
         Util.setText(input_set_buttons[c], s);
     }
+
     private void rangeButtonRefresh(final int c) {
         String lval = mMeter.getRangeLabel(chanEnum(c));
         autoButtonRefresh(range_buttons[c], lval, mMeter.getRangeAuto(chanEnum(c)));
     }
-    private void valueLabelRefresh(final int c,final MeterReading val) {
+
+    private void valueLabelRefresh(final int c, final MeterReading val) {
         final TextView v = value_labels[c];
         final String label_text = val.toString();
         Util.setText(v, label_text);
     }
+
     private void zeroButtonRefresh(final int c, MeterReading value) {
-        Log.i(TAG,"zerorefresh");
+        Timber.i("zerorefresh");
         final String s;
-        if(value.value == 0.0) {
+        if (value.value == 0.0) {
             // No offset applied
             s = "ZERO";
         } else {
@@ -376,10 +351,11 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
         }
         Util.setText(zero_buttons[c], s);
     }
+
     private void soundButtonRefresh(final int c) {
-        Log.i(TAG,"soundrefresh");
+        Timber.i("soundrefresh");
         final Button b = sound_buttons[c];
-        final String s = "SOUND:"+(mMeter.speech_on.get(chanEnum(c))?"ON":"OFF");
+        final String s = "SOUND:" + (mMeter.speech_on.get(chanEnum(c)) ? "ON" : "OFF");
         Util.setText(b, s);
     }
 
@@ -389,18 +365,14 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
 
     PopupMenu popupMenu = null;
 
-    private void makePopupMenu(List<String> options,View anchor, NotifyHandler on_choice) {
-        if(popupMenu!=null) {
+    private void makePopupMenu(List<String> options, View anchor, NotifyHandler on_choice) {
+        if (popupMenu != null) {
             return;
         }
-        popupMenu = new PopupMenu(this,anchor);
-        popupMenu = Util.generatePopupMenuWithOptions(this, options, anchor, on_choice, new Runnable() {
-            @Override
-            public void run() {
-                popupMenu = null;
-            }
-        });
+        popupMenu = new PopupMenu(this, anchor);
+        popupMenu = Util.generatePopupMenuWithOptions(this, options, anchor, on_choice, () -> popupMenu = null);
     }
+
     private void onInputSetClick(final int c) {
         final List<MooshimeterDeviceBase.InputDescriptor> l = mMeter.getInputList(chanEnum(c));
         final List<String> sl = Util.stringifyCollection(l);
@@ -408,13 +380,14 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
             @Override
             public void onReceived(double timestamp_utc, Object payload) {
                 popupMenu = null;
-                if(0!=mMeter.setInput(chanEnum(c), l.get((Integer) payload))) {
+                if (0 != mMeter.setInput(chanEnum(c), l.get((Integer) payload))) {
                     setError("Invalid input change!");
                 }
                 refreshAllControls();
             }
         });
     }
+
     private void onRangeClick(final int c) {
         List<String> options = mMeter.getRangeList(chanEnum(c));
         options.add(0, AUTORANGE);
@@ -431,36 +404,41 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
             }
         });
     }
+
     private void onSoundClick(final int c) {
-        Log.i(TAG, "onCh" + c + "SoundClick");
-        if(mMeter.speech_on.get(chanEnum(c))) {
-            mMeter.speech_on.put(chanEnum(c),false);
+        Timber.i("onCh" + c + "SoundClick");
+        if (mMeter.speech_on.get(chanEnum(c))) {
+            mMeter.speech_on.put(chanEnum(c), false);
         } else {
             mMeter.speech_on.put(chanEnum(c == 0 ? 1 : 0), false);
-            mMeter.speech_on.put(chanEnum(c),true);
+            mMeter.speech_on.put(chanEnum(c), true);
         }
         soundButtonRefresh(0);
         soundButtonRefresh(1);
     }
 
     public void onCh1InputSetClick(View v) {
-        Log.i(TAG, "onCh1InputSetClick");
+        Timber.i("onCh1InputSetClick");
         onInputSetClick(0);
     }
+
     public void onCh1RangeClick(View v) {
-        Log.i(TAG, "onCh1RangeClick");
+        Timber.i("onCh1RangeClick");
         onRangeClick(0);
     }
+
     public void onCh2InputSetClick(View v) {
-        Log.i(TAG, "onCh2InputSetClick");
+        Timber.i("onCh2InputSetClick");
         onInputSetClick(1);
     }
+
     public void onCh2RangeClick(View v) {
-        Log.i(TAG, "onCh2RangeClick");
+        Timber.i("onCh2RangeClick");
         onRangeClick(1);
     }
+
     public void onRateClick(View v) {
-        Log.i(TAG, "onRateClick");
+        Timber.i("onRateClick");
         List<String> options = mMeter.getSampleRateList();
         options.add(0, AUTORANGE);
         makePopupMenu(options, rate_button, new NotifyHandler() {
@@ -477,8 +455,9 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
             }
         });
     }
+
     public void onDepthClick(View v) {
-        Log.i(TAG, "onDepthClick");
+        Timber.i("onDepthClick");
         List<String> options = mMeter.getBufferDepthList();
         options.add(0, AUTORANGE);
         makePopupMenu(options, depth_button, new NotifyHandler() {
@@ -495,55 +474,60 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
             }
         });
     }
+
     public void onGraphClick(View v) {
-        Log.i(TAG, "onGraphClick");
+        Timber.i("onGraphClick");
         pushActivityToStack(mMeter, GraphingActivity.class);
     }
+
     public void onLoggingClick(View v) {
-        Log.i(TAG, "onLoggingClick");
-        if(mMeter.getLoggingStatus()!=0) {
+        Timber.i("onLoggingClick");
+        if (mMeter.getLoggingStatus() != 0) {
             setError(mMeter.getLoggingStatusMessage());
             mMeter.setLoggingOn(false);
         } else {
             mMeter.setLoggingOn(!mMeter.getLoggingOn());
         }
     }
+
     public void onZeroClick(final int c) {
         MooshimeterControlInterface.Channel channel = chanEnum(c);
         float offset = mMeter.getOffset(channel).value;
-        if(offset==0) {
+        if (offset == 0) {
             mMeter.setOffset(channel, mMeter.getValue(channel).value);
         } else {
             mMeter.setOffset(channel, 0);
         }
     }
-    public void onCh1ZeroClick (View v) {
-        Log.i(TAG,"onCh1ZeroClick");
+
+    public void onCh1ZeroClick(View v) {
+        Timber.i("onCh1ZeroClick");
         onZeroClick(0);
     }
-    public void onCh1SoundClick(View v) { onSoundClick(0); }
-    public void onCh2ZeroClick (View v) {
-        Log.i(TAG, "onCh2ZeroClick");
+
+    public void onCh1SoundClick(View v) {
+        onSoundClick(0);
+    }
+
+    public void onCh2ZeroClick(View v) {
+        Timber.i("onCh2ZeroClick");
         onZeroClick(1);
     }
+
     public void onCh2SoundClick(View v) {
         onSoundClick(1);
     }
+
     public void onPowerButtonClick(View v) {
-        Log.i(TAG, "onPowerButtonClick");
+        Timber.i("onPowerButtonClick");
         final List<MooshimeterDeviceBase.InputDescriptor> list = mMeter.getInputList(MooshimeterControlInterface.Channel.MATH);
         final List<String> slist = Util.stringifyCollection(list);
         makePopupMenu(slist, power_button, new NotifyHandler() {
             @Override
             public void onReceived(double timestamp_utc, Object payload) {
-                MooshimeterDeviceBase.InputDescriptor id  = list.get((Integer)payload);
-                mMeter.setInput(MooshimeterControlInterface.Channel.MATH,id);
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        mathButtonRefresh();
-                    }
-                });
+                MooshimeterDeviceBase.InputDescriptor id = list.get((Integer) payload);
+                mMeter.setInput(MooshimeterControlInterface.Channel.MATH, id);
+                runOnUiThread(() -> mathButtonRefresh());
             }
         });
     }
@@ -571,23 +555,20 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
 
     @Override
     public void onSampleReceived(double timestamp_utc, MooshimeterControlInterface.Channel c, MeterReading val) {
-        switch(c) {
+        switch (c) {
             case CH1:
             case CH2:
                 valueLabelRefresh(c.ordinal(), val);
                 // Run the autorange only on channel 2
-                if (c== MooshimeterControlInterface.Channel.CH2 && autorange_cooldown.expired) {
+                if (c == MooshimeterControlInterface.Channel.CH2 && autorange_cooldown.expired) {
                     autorange_cooldown.fire(500);
-                    Util.dispatch(new Runnable() {
-                        @Override
-                        public void run() {
-                            if(mMeter.applyAutorange()) {
-                                refreshAllControls();
-                            }
+                    Util.dispatch(() -> {
+                        if (mMeter.applyAutorange()) {
+                            refreshAllControls();
                         }
                     });
                 }
-                if(mMeter.speech_on.get(c)) {
+                if (mMeter.speech_on.get(c)) {
                     speaksOnLargeChange.decideAndSpeak(val);
                 }
                 break;
@@ -601,28 +582,34 @@ public class DeviceActivity extends MyActivity implements MooshimeterDelegate {
     public void onBufferReceived(double timestamp_utc, MooshimeterControlInterface.Channel c, float dt, float[] val) {
 
     }
+
     @Override
     public void onSampleRateChanged(int i, int sample_rate_hz) {
         rateButtonRefresh();
     }
+
     @Override
     public void onBufferDepthChanged(int i, int buffer_depth) {
         depthButtonRefresh();
     }
+
     @Override
     public void onLoggingStatusChanged(boolean on, int new_state, String message) {
         loggingButtonRefresh();
     }
+
     @Override
     public void onRangeChange(MooshimeterControlInterface.Channel c, MooshimeterDeviceBase.RangeDescriptor new_range) {
         rangeButtonRefresh(c.ordinal());
     }
+
     @Override
     public void onInputChange(MooshimeterControlInterface.Channel c, MooshimeterDeviceBase.InputDescriptor descriptor) {
         inputSetButtonRefresh(c.ordinal());
     }
+
     @Override
     public void onOffsetChange(MooshimeterControlInterface.Channel c, MeterReading offset) {
-		zeroButtonRefresh(c.ordinal(), offset);
+        zeroButtonRefresh(c.ordinal(), offset);
     }
 }
