@@ -1,21 +1,22 @@
 package com.mooshim.mooshimeter.activities;
 
-/**
- * Created by First on 2/3/2015.
- */
-
 import android.app.Application;
 
+import com.mooshim.mooshimeter.BuildConfig;
 import com.mooshim.mooshimeter.common.Util;
 import com.crashlytics.android.Crashlytics;
 import io.fabric.sdk.android.Fabric;
+import timber.log.Timber;
 
-public class MyApplication extends Application {
+public class MooshiApplication extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
         Fabric.with(this, new Crashlytics());
         Util.init(this);
+        if (BuildConfig.DEBUG) {
+            Timber.plant(new Timber.DebugTree());
+        }
     }
 }
 
